Upated billsheet and service_code tables
- Removed RecordNumber from billsheet table.
- Added Unit Rate and Description to service_code table.
- Have BilledAmount field of billsheet table automatically calculated
and updated.  It depends on the service code and the number of units
entered.

+++ server/sql/tables/unitBlock.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS `unit_block` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `consumer` int(11) NOT NULL,
   `serviceCode` int(11) NOT NULL,
-  `units` int(11) DEFAULT 0,
+  `units` float DEFAULT 0.0,
   PRIMARY KEY (`id`),
   KEY `ID` (`id`)
 --  CONSTRAINT `fkconsumer` FOREIGN KEY (`consumer`) REFERENCES `consumer` (`id`)
@@ -16,11 +16,11 @@ CREATE TABLE IF NOT EXISTS `unit_block` (
 LOCK TABLES `unit_block` WRITE;
 /*!40000 ALTER TABLE `unit_block` DISABLE KEYS */;
 INSERT INTO `unit_block` VALUES
-	(NULL, 1, 1, 111),
-	(NULL, 1, 2, 222),
-	(NULL, 1, 3, 333),
-	(NULL, 2, 2, 123),
-	(NULL, 2, 3, 456);
+	(NULL, 1, 1, 111.0),
+	(NULL, 1, 2, 222.0),
+	(NULL, 1, 3, 333.0),
+	(NULL, 2, 2, 123.0),
+	(NULL, 2, 3, 456.0);
 /*!40000 ALTER TABLE `unit_block` ENABLE KEYS */;
 UNLOCK TABLES;
